Re-structured validator code
